fix: refactor tests code
